Move coinGecko prices functions into prices.service
